refactor: S3 디렉토리 구조 변경
- 향후 확장성 고려한 디렉토리 구조 변경
